# AppointmentApi/wwwroot/swagger/v1/swagger.yaml
openapi: 3.0.3

info:
   title: Open API Spec Calendar Appointments App
   version: "1.0"
   description: This OpenAPI Specification will Display Tasks of an individual on any particular day

externalDocs:
  description: More info about Restful API
  url: https://aws.amazon.com/what-is/restful-api/#:~:text=RESTful%20API%20is%20an%20interface,applications%20to%20perform%20various%20tasks.

servers:
   - url: https://localhost:5187

tags:
   - name: Appointments
     description: Appointments on a day

paths:
  /v1/appointments:
    get:
       operationId: GetAppointments
       tags:
         - Appointments
       summary: Get appointments of the day
       parameters:
         - name: date
           in: query
           description: to get the particular date
           required: true
           schema:
             type: string
             description: to get the tasks for the date given
       responses:
         '200':
           description: Successfull -- It will return Tasks array
           content:
            application/json:
              schema:
               type: array
               items:
                 $ref: 'schemas/appointment.yaml#/components/schemas/Appointment'

         '400': 
            $ref: 'responses/common.yaml#/components/responses/BadRequest'

         '500':
           $ref: 'responses/common.yaml#/components/responses/InternalServerError'
             
     
    post:
      operationId: CreateAppointment
      tags:
      - Appointments
      summary: Create an Appointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: 'schemas/appointment.yaml#/components/schemas/AppointmentRequest'
      responses:
        '201':
          $ref: 'schemas/common.yaml#/components/schemas/GuidValueResult'
          
        '400':
          $ref: 'responses/common.yaml#/components/responses/BadRequest'

        '409':
          $ref: 'responses/common.yaml#/components/responses/ConflictError'
        
        '500':
          $ref: 'responses/common.yaml#/components/responses/InternalServerError' 
      

  /v1/appointments/{id}:
    delete:
       operationId: DeleteAppointment
       tags:
        - Appointments
       summary: Delete an appointment
       parameters:
       - name: id
         in: path
         required: true
         description: Id of the appointment to be deleted
         schema:
           type: string
           description: to update an appointment
           format: uuid
       responses:
         '204':
            description: Task successfully delete
            
         '404':
            description: appointment not found

         '500':
           $ref: 'responses/common.yaml#/components/responses/InternalServerError'
  
    